// src/app/lib/generated/models/PostCustomChatBotPayload.ts
export interface PostCustomChatBotPayload {
    /**
     * 
     * @type {ChatbotType}
     * @memberof PostCustomChatBotPayload
     */
    type: ChatbotType;
    /**
     * 
     * @type {AnimeLora}
     * @memberof PostCustomChatBotPayload
     */
    animeLora?: AnimeLora | null;
    /**
     * 
     * @type {Ethnicity}
     * @memberof PostCustomChatBotPayload
     */
    ethnicity?: Ethnicity | null;
    /**
     * 
     * @type {FantasyRace}
     * @memberof PostCustomChatBotPayload
     */
    fantasyRace?: FantasyRace | null;
    /**
     * 
     * @type {number}
     * @memberof PostCustomChatBotPayload
     */
    age: number;
    /**
     * 
     * @type {AvatarLooks}
     * @memberof PostCustomChatBotPayload
     */
    looks: AvatarLooks;
    /**
     * 
     * @type {Occupation}
     * @memberof PostCustomChatBotPayload
     */
    occupation?: Occupation | null;
    /**
     * 
     * @type {Array<Hobby>}
     * @memberof PostCustomChatBotPayload
     */
    hobbies?: Array<Hobby> | null;
    /**
     * 
     * @type {Personality}
     * @memberof PostCustomChatBotPayload
     */
    personality: Personality;
    /**
     * 
     * @type {Array<SpecialFeatures>}
     * @memberof PostCustomChatBotPayload
     */
    specialFeatures: Array<SpecialFeatures>;
    /**
     * 
     * @type {RelationshipStatus}
     * @memberof PostCustomChatBotPayload
     */
    relationshipStatus: RelationshipStatus;
    /**
     * 
     * @type {string}
     * @memberof PostCustomChatBotPayload
     */
    clientId: string;
    /**
     * 
     * @type {string}
     * @memberof PostCustomChatBotPayload
     */
    userId?: string | null;
}